requirements.txt
numpy
tqdm
opencv-python
pillow
transparent-background